feat: add orders relationship to product model

+++ oms/app/Http/Livewire/ProductsList.php
@@ -93,12 +93,25 @@ class ProductsList extends Component
 
     public function delete(Product $product): void
     {
+        if ($product->orders()->exists()) {
+            $this->addError('orderexist', 'This product cannot be deleted, it already has orders');
+            return;
+        }
+
         $product->delete();
     }
 
     public function deleteSelected(): void
     {
-        $products = Product::whereIn('id', $this->selected)->get();
+        $products = Product::with('orders')
+            ->whereIn('id', $this->selected)->get();
+
+        foreach ($products as $product) {
+            if ($product->orders()->exists()) {
+                $this->addError("orderexist", "Product <span class='font-bold'>{$product->name}</span> cannot be deleted, it already has orders");
+                return;
+            }
+        }
 
         $products->each->delete();
